feature/TC-01
- comments resolve
- create folder structure of dynamic routing
