+++ src/example/dynamic-routing/config/RouteConfig.tsx
@@ -1,25 +1,20 @@
 import React from "react";
-import { Route } from "../model/Route";
+import { Route } from "../interfaces/Route";
 import Home from "../components/Home";
 import Contact from "../components/Contact";
-import User from "../components/User";
+import About from "../components/About";
 
 const MenuItems: Array<Route> = [
-  { path: "/", title: "Home", content: <Home />, exact: true },
+  { path: "/", title: "Home", content: <Home /> },
   {
     path: "/contact",
     title: "Contact",
     content: <Contact email="haris@nisum.com" mobile="+91 99795 38985" />,
   },
   {
-    path: "/user/:userId/:abcd",
+    path: "/about",
     title: "About",
-    containParam: true,
-    params: [
-      { id: "userId", value: "1" },
-      { id: "name", value: "harry" },
-    ],
-    content: <User />,
+    content: <About />,
   },
 ];
 export default MenuItems;
