+++ src/example/dynamic-routing/interfaces/Route.ts
@@ -1,0 +1,12 @@
+
+export interface Route {
+    title: string;
+    path: string;
+    content: string | React.ReactNode;
+    hide?: boolean;
+}
+
+// export interface Param {
+//     id: string;
+//     value: string;
+// }
